requirements.txt
pip
python-dotenv
ruff
tqdm
loguru
typer
jupyterlab
numpy
torch
pandas
pillow
albumentations
scikit-learn
timm
seaborn
onnxscript
onnxruntime
gradio
ipywidgets
